Fix small issues flagged during #96 review
- CMakeLists.txt: drop /D_USE_MATH_DEFINES from MSVC branch for
  consistency with the non-MSVC drop; codebase uses its own Pi constant
  and never references M_PI.
- btAlignedAllocator.cpp: fix the second (debug-only) unsigned long →
  uintptr_t pointer-arithmetic site; same Windows LLP64 truncation bug
  as the one fixed in the preceding commit.
- run_example.sh: replace the non-existent -np flag with -pm off on
  SnowFlake and Octocat. -np was silently ignored by main's arg parser,
  so preprocess_mode was defaulting to auto instead of off.

+++ src/btConvexHull/btAlignedAllocator.cpp
@@ -112,14 +112,14 @@ void* btAlignedAllocInternal(size_t size, int32_t alignment, int32_t line, char*
 {
     void* ret;
     char* real;
-    unsigned long offset;
+    uintptr_t offset;
 
     gTotalBytesAlignedAllocs += size;
     gNumAlignedAllocs++;
 
     real = (char*)sAllocFunc(size + 2 * sizeof(void*) + (alignment - 1));
     if (real) {
-        offset = (alignment - (unsigned long)(real + 2 * sizeof(void*))) & (alignment - 1);
+        offset = (alignment - (uintptr_t)(real + 2 * sizeof(void*))) & (alignment - 1);
         ret = (void*)((real + 2 * sizeof(void*)) + offset);
         *((void**)(ret)-1) = (void*)(real);
         *((int32_t*)(ret)-2) = size;
